Integrate marketplace with NIP-90 DVM implementation
Type alignment:
- Change DVMListing.kind from u32 to u16 (NIP-90 kinds 5000-5999)
- Update get_dvm_icon() to use u16 kind constants

NIP-90 protocol in nostr-chat:
- Add PaymentRequired and Partial variants to DvmJobStatus
- Add submit_job() method to publish job requests to relays
- Add subscribe_to_dvm_results() for result/feedback subscriptions
- Add extract_amount() and extract_bolt11() helpers

Marketplace-Nostr bridge:
- Add nostr, nostr-chat, tokio dependencies to marketplace
- Create NostrBridge module for UI-to-ChatState communication
- submit_job_to_dvm(), poll_events(), active_jobs() methods

Interactive DVMList component (GPUI entity):
- Convert dvm_list.rs to proper GPUI entity with Render impl
- OnUseCallback type for USE button click handling
- merge_discovered() for mock fallback when offline

MarketplaceScreen integration:
- Add with_nostr() constructor for Nostr-enabled marketplace
- process_nostr_events() routes ChatEvents to activity feed
- Services tab renders interactive DVMList entity

Also includes TypeScript cleanup and other pending changes.

+++ crates/marketplace/src/types.rs
@@ -374,13 +374,14 @@ impl PricingUnit {
     }
 }
 
-/// A Data Vending Machine listing
+/// A Data Vending Machine listing (NIP-90)
 #[derive(Clone, Debug)]
 pub struct DVMListing {
     pub id: String,
     pub name: String,
     pub description: String,
-    pub kind: u32,
+    /// NIP-90 job kind (5000-5999)
+    pub kind: u16,
     pub provider_pubkey: String,
     pub provider_name: String,
     pub sats_per_unit: u64,
@@ -391,7 +392,7 @@ pub struct DVMListing {
 }
 
 impl DVMListing {
-    pub fn mock(name: &str, kind: u32, sats: u64, unit: PricingUnit) -> Self {
+    pub fn mock(name: &str, kind: u16, sats: u64, unit: PricingUnit) -> Self {
         Self {
             id: format!("dvm_{}", name.to_lowercase().replace(' ', "_")),
             name: name.to_string(),
